Handle subscribing to different-type-same-name exchange
If an exchange is already declared with a type in RabbitMQ, and
a client try to subscribe/declare the same exchange with different
type, RabbitMQ will close the channel and thus crash the broker.

This commit handle errors all the way from the RabbitMQ to broker
app, generating a new channel when it crashed. It also lets the
sockjs_mq know of error and not store the exchange into its dict.

This effectively stops any routing key to be bound to that exchange.

The broker will operate as normal.