Publish events when client (dis)connects
When a client connects or disconnects, the broker will publish an
event to a system exchange and other applications can listen to if
needed.

The system exchange is defined in broker/src/broker.app.src and is
defaulted to <<"private-broker">>. When client connects, the event
"connected" is sent with the payload of {socket_id}. When client
disconnects, the event "disconnected" is sent with the payload of
{socket_id, exchanges} where exchanges is a list of exchanges the
client was subscribed to.

Also, client can only publish event when in "secret-" prefixed
exchange.