Assign a queue and exchange when login
Each user after logging in will declare own exchange and bind a queue
to it. The queue will listen to every message sent to the exchange.

This is the first step to establish an exchange-to-exchange mesh.

Next up: Bind user's own exchange to followee's exchange when the user
follows somebody. That routing key will be of followee's name.