Fix unsubscription bug in Erlang broker
The broker crashed when refreshing the page because all the channels
were not closed correctly. Now let the supervisor to terminate the subscription
child instead.

Also cancel the consumer before deleting the queue to avoid unnecessary
info message. Usually, when you delete a queue without canceling the
consumer, the consumer will be canceled by the broker, but it will send out
info message.