WIP - Feed model and worker
Added JFeed and relationship to JAccount.
Switched staticemitterdelegate to koding-eventemitter2.
Added Feeder (worker) to consume and handle new activity on
'feed.new' event.

Need to do:
Set up a feed per user so that worker can query.
Bind to "FollowCountChanged" to establish exchange-to-exchange
binding so that messages will be delivered into followers' feed.