Feeder receives message from feed's owner's exchange only
Even with exchange-to-exchange-binding, the message feeder worker
receives is still from the publisher's exchange, not the follower's, thus
no way to know which feed to save the activity into.

Therefore, once the follower's queue receives messages (sent from
exchange-to-exchange), it will publish to its own exchange with a
different routing key so that the feeder can receive it.

Tested with exchange-to-exchange bindings establish from the web
management tool. Haven't figured out how to listen on following event.