Handle more negative cases in erlang broker.
* Most important, when trying to connect to an unauthorized vhost.
By design, it will return the previous connections dictionary. Then
no subscription will be made on that exchange, and thus any event
on that exchange will be ignored.
* Also not sending system events when no vhost established. This
is the case when client refreshes browser, and previous disconnected
event is fired, but no vhost at this time.
* Also check for at least one vhost connection for presence, even not
used right now.
* Terminate the broker app when the port is in used instead of leaving
it running with no connection.