Close MQ channel when user refreshes.
Do not emit "AccountIsAuthenticated" in JUser.authenticateClient, but
rely on social's worker to emit them. This is the case when account is
already logged in. JUser.login and JUser.register still emit as usual.

For each account, if the broker already handles it, the feeder will close
the channel that declared the queue and then destroy the queue so
that the next time the user open coding, previous channel is closed.

In broker module, close the channel that declared the exchange in method
#on.

As noted by Chris, dumping event into model's constructor is bad and
we need to revisit the security implication of that later.