Fix too many files opening error (EMLIMIT)
fs.readFile tries to open hundreds of files simultaneously, which is
holded by the OS file limit (that's why we are using ulimit -n to
prevent over use if file descriptors). Instead of open them all
together we push the results in a buffer with lenght of 100. This way we
are limint the # of functions executed (in our case fs.readFile)