kontrol: enable startProcess after broker is succesfull
Problem: Normally a worker starts immediately after the processes wrapper sends
a "start" command. But for processes like broker, there is a certain startup
duration that can last much longer. That means the broker restarts itself
several time until its in successful state.  However because of the restarts,
kontrol daemon has now several "start" commands and therefore adds
these for monitoring.

Solution: This merge request doesn't send a start command by default for every
restart.  It only sends if the broker is in a successful state. It makes this
via a helper function that fires after the process is successful, and therefore
prevent to overflow kontrold with "start" commands.

Additional Info: The only change on the broker code is adding a helper function
and the new -u flag (which is optional). This flag is used to share a common
uuid between the wrapper, broker and kontold. It's optional and not using the
flag doesn't break any broker functionality.