rerouting: fix a race condition and simplify code
It can happen that when rerouting starts for the first time that it receives an auth.leave instead of auth.join. If that happens the ConsumerTag which is fetched by the incoming routingKey will be empty. Thus the following functions will be not working correctly. This fixes this by checking the value of the map.