close the group channel before registering
registering will add a member to the koding group, which will in turn cause the channel to cycle, which will in turn cause the channel to be authenticated again, which will then result in a leaked guest account.  To avoid this, we should close the channel first.  I learned that after 2 days of debugging.