payment: introducing "priceIsVolatile" feature
some products do not have a fixed price.  An example of this is domain names.  For those, we need a notion of products that we can save to our database, but that are not saved as plans to recurly.  Then when we actually go to sell a product, we can provide the price at the time of the transaction.