name: explicitly call JName.one (rather than @one)
New relic magic was causing the web server to fail here.  Since we don't expect to subclass JName, this is an almost-ok solution.  We need to watch out for new relic causing us other problems via this function reapplication.  I couldn't manage to create a trivial failing case, so I'll wait to open a bug report for now.