terminal: only open 1 terminal session (per vm) when old sessions fail
This is the result of having added 1 listener per vm.  We were trying to avoid that by using a dirty flag per-vm, but that failed because of a bug where the dirty flag was getting unset whenever the state of the tabs changed, so every time a tab is closed, which happens when we cancel the terminal.  Prcatically, that meant that for N old sessions that could not be restored (per vm), we would end up adding the listener N - 1 times, e.g. if you previously had 3 tabs open, clicking "Create a new session" would open 2 new tabs.