builder: don't include versions in the file name
If we include the version in the file name, we can cause 404 errors
when we deploy a new version, because the previous verion's build
artifacts will not be available in the version's webserver.  We
could keep those artifacts around from build to build, which would
be a more optimal solution, but frankly, it constitutes more work
than this patch does.  So this is a quick fix.