ComputeProvider: create: creates a machine for given provider
This method first asks to provider to create machine with given information (to
make sure provider is ready to create that machine ~ like checking credits etc.)
when it gets the necessary information to create a JMachine document as "meta",
it uses ComputeProvider::createMachine method to create it as a mongo document.
Then when its done with the document it also calls postCreate method for given
provider which helps provider API developers to do some other operations after
machine createad as a document such as creating domains for Koding's provider.