ComputeProvider: update revive decorator to use reviveClient helper if needed
So, instead of fetching group and user from the context all the time we can use
this to get them via simple flag. Revived objects will be attached to client.r:

 client.r = { user  : JUser of client.connection.delegate,
              group : JGroup of client.context.group }

In the code that we use revive we can say:

  test = secure revive yes, (client, options, callback)->
    callback "Hello #{client.r.user.username} on #{client.r.group.title}!"

If any error occurs while fetching group or user it will call the callback
with the error.