FSHelper: createFileInstance base method for all kind of file instance creation
FSHelper.createFileInstance takes options to create a file instance for client
side file operations. If a Machine instance passed as option, it uses that to do
the operation on provided machine. If not it creates a new DummyMachine which is
not doing any file operations at all but keeps silence if one requested.