oskite: various fixes around the updater
1. Fixes a bug that was introduced last week. Mainly we weren't
updating the state to STOPPED and it was causing hostkite=null and
state=RUNNING. Basically the updateState() functin was comparing the
passed state with the LXC state. If they are the same it wouldn't update
the DB state. That's why we were passing STOPPED explicitly and LXC
would give us STOPPED, causing the state in DB not the be updated.
Instead we need to pass the DB's current state.  I've changed the
function name to avoid further confusions and fixed it by passing the
DB state for comparing.

2. There is a potential race condition during an updateState session and
an individual prepare/unprepare that would cause another DB updateState.
Basically they weren't protected, which could cause to a race condition.
I've fixed it by introducion a wait group, basically a waitgroup waits
until some work is finished. All the queues are waiting for the
waitgroup to be finished. If there is no work they continue, however if
the updaterState function starts (every 30 second), each of the queues
waits until the updater has been finished. Therefore they never cross
the line together and there is no more race condition.