kite: fix a very serious bug, details inside!
Hi :) This is just a minor bug that i've found during working on Kloud.
It's already pushed to koding/kite. Basically we would allow to
append a handler to the list of prehandlers each time we would call a
method. This fixes it it by introducting a simple bool to initialize it
just once.