activity: add scroll offsets to share popup positioning [fixes #78408034]
The amount of scrolling that has been done of the viewport area (or any
other scrollable element) is taken into account when computing the
bounding rectangle. This means that the top and left property change
their values as soon as the scrolling position changes (so their values
are relative to the viewport and not absolute).

https://developer.mozilla.org/en-US/docs/Web/API/Element.getBoundingClientRect