kloud/cloudinit: Initial implementation.
Cloud-init is an OS package that handles early initialization of a cloud
instance.

Basically, you create a YAML file that contains which commands to run,
what files to write/read etc. and it sequentially executes them, and
optionally notifies you when it is done (e.g. Sending a POST request to
a predefineda URL.)

Initial implementation has all the initialization steps that we've done
before, there should be nothing missing.

A basic VM creation timing says that we can create and SSH into a
machine within `55-57` seconds. It could be improved though.