wercker: Build and deploy klient
klient.deb and klient.gz builder is wercker now, it was jenkins before.

All build and deploy steps are removed and all wercker.yml is about
building and deploying klient.deb to S3. Why? Because:

1. Wercker does not allow to deploy a push made to a certain branch.
2. Wercker does not have a build/deploy filter.

So, this branch's wercker.yml differed from other branches. When a push
is made to this branch, wercker will be notified and only klient will be
deployed, nothing else.

Same actions will take effect for klientkite.production branch.