kloud/cloudinit: Put the migrate.sh in its place.
This was a regression that took my 3 hours to bisect and debug.
I'm having a hate relationship with YAML now thanks to this bug.

The main bug is in the first line of README.md. It contains a # and a
whitespace character and YAML interprets it as a comment since it is the
first line of a block quote. Yaml implementation should interpret as it
is a block quote, not a comment line. Stupid implementation.

The second one is the `owner` parameter of the `write_files` directive.
It is broken somehow, I know the exception but I'm really tired of this
shit.