kloud/cloudinit: Retry on connrefused errors.
Some users' VMs are automatically shutdown although the VM is actively
used. The problem was that klient was not installed on their system.

S3 is refusing to accept any incoming connections. We don't know when
and there is no way to debug the problem. So while installing the klient
package from S3, treat the `connection
refused` errors as transient and retry 5 times on transient
errors.