viewport: substitute `document.body.scrollTop` with `window.scrollY` [#80403198]
How `document.body.scrollTop` property is set specified in the CSSOM
View Module specification. [1] It looks like specification is missing
how user agent should set the property when it is not in quirks
mode. Therefore, user agents apply sixth step about setting the
`scrollTop` attribute of `document.body`. Basically `scrollTop` property
is not set.

Blink, Chromium's rendering engine, is applying second step of getting
`scrollTop` attribute of `document.body` element because when `document`
is not in standards mode and thus skipping the first step.

Seems like other browsers are not returning the value of
`window.scrollY` when getter of `document.body.scrollTop` property is
invoked in standards mode.

Also, I have found a mailing list thread regarding this subject. One
message [2] is pointing out specification is not covering behavior of
`document.body.scrollTop` when user agent is not in quirks mode.

[1] http://dev.w3.org/csswg/cssom-view/#dom-element-scrolltop
[2] http://lists.w3.org/Archives/Public/www-style/2013Oct/0287.html

Signed-off-by: Sonmez Kartal <sonmez@koding.com>