kloud: Ping the klient to check the machine state.
Just checking Kontrol if the klient is registered leads to a bad UX
because when a klient is disconnected, Kontrol still shows it as
registered before a certain timeout.

We now ping the klient to see if it is running. In other cases, we just
ask AWS about the machine state again.

Client side is calling klient.ping to see if it is running instead of
asking kloud now. Hopefully these changes will improve the current
rate-limit situation.