kloud: Reduce state poller interval from waitstate.
Most of the DescribeInstances calls are made from the eventer itself as
it turns out. Separate the poller and eventer ticker intervals. We're
gonna poll less (30 seconds by default) and push events more (3 seconds
by default).

This change will affect the UI where the state changes are not be
visible instantly but will reduce the overall system stability.