kloud/waitstate: lazy finish for terminating
We don't need to wait until we have the final terminated state. Once the machine is in "terminating" state there is no going back anymore. So after that state we can assume the machine is terminated and move one. This is improves the overall stability, lowers the total API calls and and decreases the total destroy duration.