kloud: include Amazon Support recommendations
Original e-mail from them:
-------------------------

Hello,

To give you an update from the development team, they are still
investigating the exact root cause of the issue.

They have suggested some possible workarounds to get you out of
immediate pain:

Is it possible to:
1) We understand that this is against what we generally recommend   but
is it possible for you to make the RunInstances calls from the root
account and not an IAM user.
2) Could you not specify the snapshot (the same one as the AMI) on the
block device mappings for the RunInstances calls.

Please do let us know if this is possible as the development team
continues to investigate this.