kloud: Fix wrong return value on locker.
Return the correct error value if the machine document is not found.
We had a case when a jMachine and jDomainAliases documents are in an
inconsistent state, and the user tries to set a domain. She can't
because the relation between the jDomainAliases.MachineId and
jMachines.Id are broken.

Try to recover from that state if we can't find the document and do the
job instead of failing.