klient: Add persistency to collaboration.
We use `boltdb` for storage and persist the users at
/opt/kite/klient/klient.db which is the klient's data directory.

Bolt is a fast key-value storage, it can be replaced with another
database, such as Sqlite, if necessary.

The only drawback of boltdb as far as I see is, you can't read the db if
another process has a write lock on it. Such as, if klient owns the db,
another process can't make any query on it, even if the query is
read-only.