JMachine::approve: allow users to mark their state as approved in this.users
When one shares their machine with someone else, we are adding them into users
list as approved: False which then they need to set this field to True for their
accounts if they are willing to use this machine. This method will set the
approved field as True for them after making required checks.