collaboration: add Option struct to users
This is needed so we can distinguis between the permanent mode and
non permanent mode. Because sharing on Koding can be done in two ways,
one is a temporary sharing which is caused by the collaboration part,
and it's only valid if the sharer is online (so the user on the other
end is not permanent).

Another option is to fully share their VM. Which is the permanent mode.