klient: add registration to Kontrol via Password
This PR enables the user to create a kite.key automatically by
authenticating to Kontrol directly with their Koding usernames and
password. If successfull a kite.key is created and places at
~/.kite/kite.key  By default it's disabled.

Usage is in the form of:

klient -register
klient -register -username arslan
klient -register -username arslan -kontrol-url "https://foo.company.com"

This is be used to register any third party machine to kontrol.