ManagedProvider: instead of full queryString keep only kiteId
If we keep all the query information with in queryString it will cause some
issues in future. For example if we keep version information with in JMachine
and if the klient in that machine updates it self, we will fail to connect to it
But instead we can keep continue using kiteId only which is unique. And wont
change with klient kite updates.