ComputeProvider: move default stack creation to JUser's convert step
Since the new IDE router and environment data provider is faster than before
it's possible for nusers to see No machine found state just after registration.

To prevent that we moved default stack creation into register steps so we will
be sure that machines will be there when registration completed.