terminal: use process owner, instead of the requester
Sessions are created by the process owner. So for example they can be
started by the user "arslan", if the request comes from the user
"arslan" everything is ok.

However the session can also be started as "root". If that it's the
case, the requester "arslan" is not possible to kill the sessions,
because killSessions tries to delete sessions belonging to the user
"arslan". Always use the process owner user.