Sidebar: Don't draw sidebar when account added to channel.
We are now handling sidebar redrawing with a different way which is lead by computecontroller.reset. This line was the reason of multiple machine share popups. I tested it widely it's not the case anymore but I am not sure removing this line is 100% sure. So I will leave it as commented for a while. If we start to show missing or not removing machines on sidebar when collaboration started or ended this line might be the reason. If you blame this line and see my name