kloud: improve klient checking
Problem:
If a klient didn't yet register to Kontrol, we were saying that it's ok
and would return non error state for the methods build, start and
resize. On the client side, the user would see that VM is ready, bu in
fact it's not. This means even if the VM is running, because there is
no Klient running the User would see a blank terminal in those cases.

Solution:
This PR fixes this state, we no longer mark it as healthy if we can't
reach it. Also we increase the time to check the existence from
Kontrol, this is especially an improvement in development environments
when there is an high latency in ngrok or if it's a slow AWS instance
that it takes long to get in a stable state.