kloud: improve state handling in various situations
1.
Problem:
Unmarshalling would return an error always in Info method.

Solution
Fix unmarshalling into non pointer struct by passing the address.
Because client side is handling Info errors itself, it was shadowing
this error and we wouldn't see it.

2.
Problem:
If there is no instance available during Start method we would mark it
as NotInitialized. But the Start method would return a nil error. If we
don't return an error, the Kloud controller (the dispatcher for all
methods) is returning an OK state to the eventer. That's why even if we
mark the state as NotInitialized, on the UI it would be seen as Running
and a blank screen. The user was required to refresh the page, so it
would see the NotInitialized state (aka Turn On button). This is a case
only introduced by human errors or non Kloud interaction.

Solution:
Return an error if the instance is not avaialable anymore in Start
method. Now the eventer is returning an error and the UI can act
correctly. Also we change the order of the defer to avoid double update
query to MongoDB.

3.
Problem:
If an AWS machine is destroyed without calling Kloud's destroy
method, the machine document would be still available. If someone calls
Info, it would fetch the state Terminated (because the machine is still
there). And return that particular state. The UI shows a "Creat VM"
modal if it's receives a Terminated state. But because the Machine
document is still there we would receive a "Free VM's can't create more
than one VM". This is a very rare case only introduced by human errors
or non Kloud interaction.

Solution:
We fix it by marking the machine document as NotInitialized (self heal)
so the UI side can immediately start to build.