kloud/controller: pass certain errors to client side
We have some errors we want to pass to the client side, such as Network
Limit or plan expired. We are usually not passing anything to avoid
leaking errors to client side.

We use a a new `error` type called `NewEventerError` which is used to
pass errors explicitly to the client side. We are using it to pass
certain errors from the start method to the client side.