kdbin: Added client methods, to support vm -> ui communcation
Two methods have been introduced to Klient. `client.Publish` and
`client.Subscribe`. The purpose / implementation of these methods
are as follows:

The Publish method takes arbitrary event data, and passes it to
any functions which have subscribed via `client.Subscribe`. The
only required value is a single `eventName` value. Only listeners
of the given eventName will be called back with the data.

The Subscribe method subscribes a function to any `client.Publish`
calls with the matching eventName. On client disconnect, the function
is removed.

With these generic methods, an authorized process is able to connect
to klient and send/recieve arbitrary data - though it is intended to
take data from the VM and send it to the Koding browser UI.