klient: Moved bindKlientEvents call to handle more use cases.
A first bindKlientEvents is called if the machine is already running,
and no machine state modal needs to be created. If however, the machine
is not loaded (even if it's on), a machineStateModal is loaded which
handles all of that. When it's done, it calls handleIDEBecameReady, and
as part of that bindKlientEvents.

This way we will resubscribe anytime we disconnect.