kloud: Machine.Build() fallback to base image, if snapshot doesn't exist
When building, Build() will check if the snapshot specified in
`Meta.SnapshotId` actually exists. If it does not, it will fallback to
the Koding base image[1].

[1] With exception. If a user supplied the snapshotId argument to Build,
and that Snapshot does not Exist, Build returns an error rather than
falling back to the base image. This ensures that a failure happens if
you explicitly tell Build() to do something it cannot fulfill.