managed: Fixed a couple bugs, added "UX" improvements
Bugfix: If klient was not pre-existing, the service restart would cause
an error message to print to the end user. Roughly `stop: unknown
instance`. This now pipes to devnull. (it came from stderr)

Bugfix: Because the managed install is being piped from curl, and into
bash, Klient was unable to get the correct tty, and therefor any prompts
from the user would cause a failure. For example, the replace kite key
prompt.

UX: General messages were added before any long running process.

UX: A newline is shown before running klient register, so that if it
prompts the user, the user can easily read/respond.

UX: A success message was added to the end of the script. Though, the
wording will need to be improved i'm sure.