info: Added whois to ProviderChecker type
Since the majority of the checkers will be using whois currently, it
seems best to do this. Since this primarily serves to reduce repeated
http queries to the whois server, in the future a Cache type may be best
implemented.

We could also put all of the checkers under a struct, which serves to
cache information in the struct - but i imagine keeping the checkers as
independant functions is the most appropriate design.