flux: move store registrations into maincontroller
The reason behind this change is to be able to register stores
before anything is loaded. Before this, store registrations were
happening on app controllers' constructor, which is too late for some
parts of the application. For example, if you reload on IDE, other apps
stores will not be registered, result of this sidebar will be empty.