kloud: Adjusted plan timeout to 50m
As a Cost Control measure, a 50m timeout means that we should only be
billed for 1 hour blocks, rather than 2 hour blocks.

As an example. Previously we were shutting the machine down at roughly
60m. If we went over by a tiny bit, say 61m before we shutdown, or if
we took too long to shut down, we would end up being billed for 2 hours
instead of a single hour - because Amazon bills by the hour. This
commit fixes that.