kloud: Stop VM if Klient will not connect
As part of the normal Klient timeout vm stoppage, Kloud now tracks
whether the machine's Klient is not connecting properly, and how long
that has been. If it has been longer than X minutes, Kloud stops the VM.
This helps ensure that a user can't simply stop Klient, and then get a
Free VM for until Cleaner catches it.