install: Added timeout to aws api call
To improve UX on systems where no api is available, the api call
now has a timeout of 5 seconds. Meaning that it will wait for the api
call for 5 seconds, before giving up.

This also means that the AWS api must return within 5 seconds, but in my
tests the calls are nearly instant. I believe the api calls are locally
networked, so very, very fast.