ComputeController:update: use stack template credential if possible
For teams context we need to use JMachine.credential field for ongoing
operations with ComputeProvider which will need to verify credential
on each request. There is one user experience issue with that behaivour
which is causing user to reinitialize their stacks if one of the valid
credential has been changed. To prevent that we are taking stack template
credential if it fits with the current JMachine requirements. So user
not requires to re-init their stacks when a credential is changed but not
the template itself