kloud: Do not stop alwayson machines after info decision
Note that we're chosing to still attempt the info decision (rather than
always returning on), because this helps us log the issue and keep stats
on it. Ideally, always-on machines should never return an info decision
of stopped, so it's worth while to log the issue, and be aware of
potential problematic patterns.