kloud: Log information about initial timed out aws build instance
To track meaningful metrics for the AWS build timeout investigation, we
are logging special data in the event of a timeout. We are only logging
this if we first created the instance _(ie, `Machine.Meta.InstanceId`
was empty)_, because we are only interested in the first attempts.

In addition to the special log, we also log the execution time where we
waited for the build specifically. This should normally match the
WaitState timeout time, but regardless, we're logging it for good
measure.