kloud: Handle build state recovery
Previously the build method would assume that a machine is transitioning
from Starting to Started. It would not check for existing states. This
led to cases where if the machine was in a different state, such as
stopped, the build method would just continually wait for build to
finish.. which would never happen. This is especially a problem, as
various vm cleaner workers would stop machines under certain conditions
(long running, prolonged state mismatch, etc), and a returning user
could never finish building their machine.

To resolve this, a new private method has been introduced,
`buildRecovery`. This is called in the event that an instanceId already
exists, and just serves to correct any assumptions that the rest of the
build method has (such as waiting for running). Currently, it *just*
checks for a Stopped machine, and starts it.