kloud: Assign Elastic IP before machine has started
The old method of assigning an elastic ip after the machine has started
can cause a race condition for Klient, where the machine's IP changes
moments after it is first brought up.

To resolve this, the elastic IP is assigned *before* the machine is
started (when it is in the Stopping or Stopped state).

In addition to this, the IP of the machine is no longer removed on
machine Stop. This allows the Start method to know the previous IP,
and check if it was an elastic IP or not, allocating a new one as
needed.