klientctl: Require sudo for stop/start/install/uninstall/update
klientctl is now capable of checking if it has sudo permission,
and informing the user where it is required.

Note that this commit has some rather verbose code, but it has been
designed in such a way that we can create cross-platform
implementations and also making it as testable as possible, regardless
of what a specific platforms implementation might require.