fs: Increase the klient filesize limit to 50MB
To accomidate the needs of fuseklient, the readFile limit in Klient is
being increased to 50MB. This is quite large, but the UX suffers greatly
when limited to smaller filesizes.

In the future we should reduce this number, and instead provide a
Streaming file option (for unlimited file sizes). However, that is a
core Kite API change, so this is a temporary solution to mitigate bad
UX.