ComputeController: reinitGroupStack: Reinit's given stack or groups default stack
If stack given, it asks for re-init and first deletes and then calls createDefaultStack again.
If not given it tries to find default one and does the same thing, if it can't find the default one, asks to user what to do next.

We will need this for putting rebuild button into stack template editor #106403276