JComputeStack: destroy: instead of removing mark it as Destroying [completes #104007780]
Instead of deleting the stack we let kloud to do it when destroy/reinit requested. But while waiting kloud to finish it's job we should allow user to create new stacks, so we are marking it as Destroying and removing user's from Machine users so they won't able to reach it until kloud finishes the destroy.

We are also updating counters based on this action so user plan usages will be updated as well.