status: Run status on list/mount/unmount connection related failures
Note that this is *not* bulletproof. Status checks basic things, but it
does not actively communicate with Kontrol _(such as asking Kontrol to
list kites)_, and therefor status might say everything is working fine
despite not being able to GetKites.

This is due to Kite's internal reconnect mechanism, and how long that
may take to reconnect. Backoffs, to prevent DDOS, especially hinder
this.