scripts/sl: Softlayer helper tool
The official slcli lacks couple of features like tagging the keys, which
is going to be needed in future for the packer plugin to lookup the key
to attach to the instance.

This tool is going to be extended also with more debug / maintanance
commands for the Softlayer VMs.

Moreover there're are some minor quirks in the official tool, like
the  sshkey resource refuse to cooperate, in particular the following command:

	$ slcli sshkey add -f kloud_rsa.pub -k kloud_rsa.pem kloud

Fails with:

	SoftLayerAPIError(SoftLayer_Exception_Public): Unable to generate a
	fingerprint. Please ensure the key provided is a valid public key.