restart: Always stop during restart
In the event of klient having problems, klientctl may incorrectly
determine that klient is stopped. Previously, if this happens, the
`kd restart` command would not attempt to stop Klient. With this fix, we
always stop during restart, but only notify the user of stop-errors
if we know Klient is running.