ComputeProvider: updateTeamCounters: syncs and updates the counter with real data
This one is fetches existing counts, real stack/machine counts and updates counters with the new ones. We will going to use to update existing teams usages and we will use this to repair somehow broken counters, we can use this one safely on team login or in a worker in future.