klient: Added PublicIPRetry, and PublicIP host rotation
The 2nd host and host rotation helps to ensure that if one of
the IP providers _(even us, eventually)_  is down, Klient can still
start and potentially run normally.

This does not "fix" the panic, due to the panic being intended.
The panic behavior (Seen here:
https://github.com/koding/koding/blob/master/go/src/koding/klient/app/klient.go#L386)
Is not inherently bad. Klient cannot do much if it cannot get its IP,
and if it is failing to even communicate, restarting (after a panic)
is a reasonable response.

This PR mainly focuses on the reliance to echoip, and a single provider.