klientctl: Added klientrunningrepair
Committing for historical sake, but the runningrepair is likely
going to have to be changed to use a sub process, spawned with
sudo.

The reason being, kd repair cannot use sudo, but yet service
interactions need sudo. So, we need to try making klientrunningrepair
run a subprocess of `sudo kd restart` to resolve the issue.