client: test: waitForTextContain coffee version added cc/ @fatihacet @didemacet
since I have other commands written in coffee, I need to compile this and put in the build/lib folder I made a js2coffee conversion. Probably it is a good idea to put the original coffee file here, if any.

Or, alternatively I found a coffee version of the almost same thing here PTAL:
https://github.com/maxgalbu/nightwatch-custom-commands-assertions/blob/master/coffee/commands/waitForText.coffee