tunnelproxy, klient: local route support
This CL adds local route support. The flow begins on the klient side,
when klient requires a tunnel to connect to kontrol. The tunnel
registration routine was extended with LocalRoutes parameters, which,
when set, tells tunnelserver which HTTP requests it can redirect
locally.

If the klient is inside a VirtualBox machine, the guest klient asks
host klient whether it has any forwarded ports to localhost for the
guest kite server, and if positive - it's used for the local route.

For websocket and tcp connections clients need to first call the guest
kite asking routes the tunnel.route method, and if local route is possible,
the received address should be used for connecting, instead of public
route. Task for adding support for it to client - TMS-2676.