tunnelserver: rework /resolve method -> /-/discover
Makes it possible to resolve dynamic addresses for the service, e.g.

In order to get the host for the klient kite service:

$ curl 2ba66ea1adec.rjeczalik.koding.me/-/discover/kite
2ba66ea1adec.rjeczalik.koding.me

If the request is being made from the same host that the klient kite,
local address is going to be returned instead:

$ curl 2ba66ea1adec.rjeczalik.koding.me/-/discover/kite
127.0.0.1:56789

Additionally in order to make it possible to work with the IDE, a kites
service exists on local (but it is not tunnelled, as TLS sessions are
terminated on {dev,prod}proxy instance). A "kites" serivice is basically
a kite server behind a TLS proxy. To obtain address for it:

$ curl 2ba66ea1adec.rjeczalik.koding.me/-/discover/kites
127.0.0.1:56790