klient/vagrant: ensure vagrant working dir is not dirty
If stack reinit or destroy was called when host klient (the
provisionner) was down, the destroy operation ends with failure
while leaving dirty vagrant working directory.

If no dirty directory was explicitely requested via the request, this CL
makes the klient/vagrant destroy the box prior to create if it already
exists.