ComputeController: checkMachinePermissions: ignore managed vms
we are not able to change user data on a managed vm and for that reason we are not moving ownership of kicked user managed stack, but we are moving ownership of all machines without checking it's provider. and once we kick a user the ui side is fetching all machines and stacks, ignores managed stacks of kicked user but still processes managed vms and with that process we should ignore managed ones.